+++ SmallBeachFarm/Framework/GenericModConfigMenuIntegrationForSmallBeachFarm.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Pathoschild.Stardew.Common.Integrations.GenericModConfigMenu;
 using Pathoschild.Stardew.SmallBeachFarm.Framework.Config;
 using StardewModdingAPI;
+using StardewValley;
 
 namespace Pathoschild.Stardew.SmallBeachFarm.Framework
 {
@@ -16,8 +17,8 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
         /// <summary>The Generic Mod Config Menu integration.</summary>
         private readonly GenericModConfigMenuIntegration<ModConfig> ConfigMenu;
 
-        /// <summary>Get the possible farm choices.</summary>
-        private readonly IDictionary<int, string> FarmChoices = new Dictionary<int, string>();
+        /// <summary>The IDs and display names for each farm type that can be replaced.</summary>
+        private readonly IDictionary<int, string> FarmChoices;
 
 
         /*********
@@ -27,15 +28,13 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
         /// <param name="modRegistry">An API for fetching metadata about loaded mods.</param>
         /// <param name="monitor">Encapsulates monitoring and logging.</param>
         /// <param name="manifest">The mod manifest.</param>
-        /// <param name="modData">The mod's hardcoded data.</param>
         /// <param name="getConfig">Get the current config model.</param>
         /// <param name="reset">Reset the config model to the default values.</param>
         /// <param name="saveAndApply">Save and apply the current config model.</param>
-        public GenericModConfigMenuIntegrationForSmallBeachFarm(IModRegistry modRegistry, IMonitor monitor, IManifest manifest, ModData modData, Func<ModConfig> getConfig, Action reset, Action saveAndApply)
+        public GenericModConfigMenuIntegrationForSmallBeachFarm(IModRegistry modRegistry, IMonitor monitor, IManifest manifest, Func<ModConfig> getConfig, Action reset, Action saveAndApply)
         {
             this.ConfigMenu = new GenericModConfigMenuIntegration<ModConfig>(modRegistry, monitor, manifest, getConfig, reset, saveAndApply);
-            foreach (ModFarmMapData farm in modData.FarmMaps)
-                this.FarmChoices[farm.ID] = farm.Name ?? farm.Map;
+            this.FarmChoices = this.GetFarmNamesById();
         }
 
         /// <summary>Register the config menu if available.</summary>
@@ -75,6 +74,30 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
                     set: (config, value) => config.ReplaceFarmID = this.GetFarmID(value, defaultValue: new ModConfig().ReplaceFarmID),
                     choices: this.FarmChoices.Values.OrderBy(p => p).ToArray()
                 );
+        }
+
+        /// <summary>Get the IDs and display names for each farm type that can be replaced.</summary>
+        private IDictionary<int, string> GetFarmNamesById()
+        {
+            Dictionary<int, string> farmNamesById = new();
+
+            foreach (int id in new[] { Farm.default_layout, Farm.riverlands_layout, Farm.forest_layout, Farm.mountains_layout, Farm.combat_layout, Farm.fourCorners_layout, Farm.beach_layout })
+            {
+                string translationKey = id switch
+                {
+                    Farm.default_layout => "Character_FarmStandard",
+                    Farm.riverlands_layout => "Character_FarmFishing",
+                    Farm.forest_layout => "Character_FarmForaging",
+                    Farm.mountains_layout => "Character_FarmMining",
+                    Farm.combat_layout => "Character_FarmCombat",
+                    Farm.fourCorners_layout => "Character_FarmFourCorners",
+                    Farm.beach_layout => "Character_FarmBeach",
+                    _ => throw new InvalidOperationException($"Unexpected farm ID {id}.")
+                };
+
+                farmNamesById[id] = Game1.content.LoadString(@$"Strings\UI:{translationKey}").Split('_')[0];
+            }
+            return farmNamesById;
         }
 
         /// <summary>Get the name of a farm type.</summary>