+++ SmallBeachFarm/ModEntry.cs
@@ -66,7 +66,7 @@ namespace Pathoschild.Stardew.SmallBeachFarm
 
             // read config
             this.Config = this.Helper.ReadConfig<ModConfig>();
-            this.FarmMapAssetName = this.Data.FarmMaps.FirstOrDefault(p => p.ID == this.Config.ReplaceFarmID)?.Map;
+            this.FarmMapAssetName = Farm.getMapNameFromTypeInt(this.Config.ReplaceFarmID);
             if (this.FarmMapAssetName == null)
             {
                 this.Monitor.Log("You have an invalid farm ID in the 'config.json' file. You can delete the file to reset it. This mod will be disabled.", LogLevel.Error);
@@ -174,7 +174,6 @@ namespace Pathoschild.Stardew.SmallBeachFarm
             // add Generic Mod Config Menu integration
             new GenericModConfigMenuIntegrationForSmallBeachFarm(
                 getConfig: () => this.Config,
-                modData: this.Data,
                 reset: () =>
                 {
                     this.Config = new ModConfig();